[FLINK-21649][state/heap] Extract StateMapEntry class
Motivation: ease refactorings and tracking code changes.
