+++ flink-runtime/src/main/java/org/apache/flink/runtime/state/heap/CopyOnWriteStateMap.java
@@ -30,14 +30,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.ConcurrentModificationException;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Set;
 import java.util.TreeSet;
 import java.util.stream.Stream;
@@ -802,130 +800,6 @@ public class CopyOnWriteStateMap<K, N, S> extends StateMap<K, N, S> {
 
     // StateMapEntry
     // -------------------------------------------------------------------------------------------------
-
-    /**
-     * One entry in the {@link CopyOnWriteStateMap}. This is a triplet of key, namespace, and state.
-     * Thereby, key and namespace together serve as a composite key for the state. This class also
-     * contains some management meta data for copy-on-write, a pointer to link other {@link
-     * StateMapEntry}s to a list, and cached hash code.
-     *
-     * @param <K> type of key.
-     * @param <N> type of namespace.
-     * @param <S> type of state.
-     */
-    @VisibleForTesting
-    protected static class StateMapEntry<K, N, S> implements StateEntry<K, N, S> {
-
-        /** The key. Assumed to be immumap and not null. */
-        @Nonnull final K key;
-
-        /** The namespace. Assumed to be immumap and not null. */
-        @Nonnull final N namespace;
-
-        /**
-         * The state. This is not final to allow exchanging the object for copy-on-write. Can be
-         * null.
-         */
-        @Nullable S state;
-
-        /**
-         * Link to another {@link StateMapEntry}. This is used to resolve collisions in the {@link
-         * CopyOnWriteStateMap} through chaining.
-         */
-        @Nullable StateMapEntry<K, N, S> next;
-
-        /**
-         * The version of this {@link StateMapEntry}. This is meta data for copy-on-write of the map
-         * structure.
-         */
-        int entryVersion;
-
-        /**
-         * The version of the state object in this entry. This is meta data for copy-on-write of the
-         * state object itself.
-         */
-        int stateVersion;
-
-        /** The computed secondary hash for the composite of key and namespace. */
-        final int hash;
-
-        StateMapEntry(StateMapEntry<K, N, S> other, int entryVersion) {
-            this(
-                    other.key,
-                    other.namespace,
-                    other.state,
-                    other.hash,
-                    other.next,
-                    entryVersion,
-                    other.stateVersion);
-        }
-
-        StateMapEntry(
-                @Nonnull K key,
-                @Nonnull N namespace,
-                @Nullable S state,
-                int hash,
-                @Nullable StateMapEntry<K, N, S> next,
-                int entryVersion,
-                int stateVersion) {
-            this.key = key;
-            this.namespace = namespace;
-            this.hash = hash;
-            this.next = next;
-            this.entryVersion = entryVersion;
-            this.state = state;
-            this.stateVersion = stateVersion;
-        }
-
-        public final void setState(@Nullable S value, int mapVersion) {
-            // naturally, we can update the state version every time we replace the old state with a
-            // different object
-            if (value != state) {
-                this.state = value;
-                this.stateVersion = mapVersion;
-            }
-        }
-
-        @Nonnull
-        @Override
-        public K getKey() {
-            return key;
-        }
-
-        @Nonnull
-        @Override
-        public N getNamespace() {
-            return namespace;
-        }
-
-        @Nullable
-        @Override
-        public S getState() {
-            return state;
-        }
-
-        @Override
-        public final boolean equals(Object o) {
-            if (!(o instanceof CopyOnWriteStateMap.StateMapEntry)) {
-                return false;
-            }
-
-            StateEntry<?, ?, ?> e = (StateEntry<?, ?, ?>) o;
-            return e.getKey().equals(key)
-                    && e.getNamespace().equals(namespace)
-                    && Objects.equals(e.getState(), state);
-        }
-
-        @Override
-        public final int hashCode() {
-            return (key.hashCode() ^ namespace.hashCode()) ^ Objects.hashCode(state);
-        }
-
-        @Override
-        public final String toString() {
-            return "(" + key + "|" + namespace + ")=" + state;
-        }
-    }
 
     // For testing
     // ----------------------------------------------------------------------------------------------------
